requirements.txt
addict==2.4.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
ConfigArgParse==1.7.1
contourpy==1.3.0
cycler==0.12.1
dash==3.2.0
fast_simplification==0.1.12
fastjsonschema==2.21.2
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.1
freetype-py==2.5.1
fsspec==2025.9.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.8.1
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
narwhals==2.9.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
open3d==0.18.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.1
pyglet==1.5.27
PyOpenGL==3.1.0
pyparsing==3.2.5
pyquaternion==0.9.9
pyrender==0.1.45
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
retrying==1.4.2
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
trimesh==4.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
